% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.df.df.R
\name{merge.df.df}
\alias{merge.df.df}
\title{Merge two data frames containing some common samples in rows}
\usage{
\method{merge}{df.df}(df1, df2, key)
}
\arguments{
\item{df1}{Data frame 1}

\item{df2}{Data frame 2}

\item{key}{name of column shared between df1 and df2 that has unique values and can be used to synchronize position and existence of rows in output.}
}
\description{
Merge two data frames containing some common samples in rows
}
\examples{
d1 <- data.frame(sample = paste0("PG", 1:6),
geno = sample(c("wt", "het", "ko"), 6, replace = TRUE),
batch = c(1,1,2,2,3,3))
d2 <- cbind.data.frame(d1[c(1,3,6,4,2),],
patho = sample(c("high", "med", "low"), 5, replace = TRUE))
d1.2 <- merge.df.df(df1 = d1, df2 = d2, key = "sample")

}
