% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.df.df.R
\name{sync.df.df}
\alias{sync.df.df}
\title{Sync dataframe1 and dataframe2}
\usage{
sync.df.df(df1, df2, key)
}
\arguments{
\item{df1}{dataframe1}

\item{df2}{dataframe2}

\item{key}{name of column shared between df1 and df2 that has unique values and can be used to synchronize position and existence of rows in output.}
}
\value{
A list object with synchronized data frames
}
\description{
This function simply takes in 2 separate data frames (with samples in rows and measures in columns), as input and synchronizes the rows such that the samples in both are synchronized. The function uses a key column (i.e. sampleID), that is common to both and denotes unique IDs for each row.
}
\examples{
d1 <- data.frame(sample = paste0("PG", 1:6),
geno = sample(c("wt", "het", "ko"), 6, replace = TRUE),
batch = c(1,1,2,2,3,3))
d2 <- cbind.data.frame(d1[c(1,3,6,4,2),],
patho = sample(c("high", "med", "low"), 5, replace = TRUE))
d1.2 <- sync.df.df(df1 = d1, df2 = d2, key = "sample")
}
